update README; bugs for tpl

+++ bin/bnorth.js
@@ -10,51 +10,42 @@
 
 'use strict';
 
-var spawn = require('cross-spawn');
-var script = process.argv[2];
-var args = process.argv.slice(3);
+let spawn = require('cross-spawn');
+let script = process.argv[2];
+let args = process.argv.slice(3);
+let fs = require('fs-extra');
+let chalk = require('chalk');
 
-switch (script) {
-case 'init':
 
-case 'build':
-case 'start':
-case 'test':
-case 'eject':
-
-case 'page':
-case 'component':
-
-case 'plugin':
-case 'appprepare':
-case 'appbuild':
-case 'publish':
-  var result = spawn.sync(
-    'node',
-    [require.resolve('../scripts/' + script)].concat(args),
-    {stdio: 'inherit'}
-  );
-  if (result.signal) {
-    if (result.signal === 'SIGKILL') {
-      console.log(
-        'The build failed because the process exited too early. ' +
-        'This probably means the system ran out of memory or someone called ' +
-        '`kill -9` on the process.'
-      );
-    } else if (result.signal === 'SIGTERM') {
-      console.log(
-        'The build failed because the process exited too early. ' +
-        'Someone might have called `kill` or `killall`, or the system could ' +
-        'be shutting down.'
-      );
-    }
-    process.exit(1);
-  }
-  process.exit(result.status);
-  break;
-default:
-  console.log('Unknown script "' + script + '".');
-  console.log('Perhaps you need to update react-scripts?');
-  console.log('See: https://github.com/facebookincubator/create-react-app/blob/master/packages/react-scripts/template/README.md#updating-to-new-releases');
-  break;
+let scriptPath = `../scripts/${script}` 
+if(!fs.existsSync(require.resolve(scriptPath))){
+  console.log(`Unknown command '${script}'`);
+  console.log(`Update or see ${chalk.cyan('https://github.com/able99/bnorth/blob/master/README.md')}`);
+  return;
 }
+
+
+let result = spawn.sync(
+  'node',
+  [require.resolve(scriptPath)].concat(args),
+  {stdio: 'inherit'}
+);
+
+if (result.signal) {
+  if (result.signal === 'SIGKILL') {
+    console.log(
+      'The build failed because the process exited too early. ' +
+      'This probably means the system ran out of memory or someone called ' +
+      '`kill -9` on the process.'
+    );
+  } else if (result.signal === 'SIGTERM') {
+    console.log(
+      'The build failed because the process exited too early. ' +
+      'Someone might have called `kill` or `killall`, or the system could ' +
+      'be shutting down.'
+    );
+  }
+  process.exit(1);
+}
+process.exit(result.status);
+
